+++ Exercise1/src/array.h
@@ -6,7 +6,7 @@ Array_Struct *array_create(void);
 
 void array_free_memory(Array_Struct *array_struct);
 
-int array_size(Array_Struct *array_struct);
+unsigned long array_size(Array_Struct *array_struct);
 
 void array_struct_add(Array_Struct *array_struct, void* element);
